requirements.txt
flake8
python-dotenv
psycopg2-binary
statx